* arts-2-3-879
* src/gas_abs_lookup.cc (find_new_grid_in_old_grid):

Replaced fixed tolerance of 1 Hz in comparison by a tolerance that uses
a DBL_EPSILON times the frequency value being compared. This should
prevent two neighboring frequencies being so close together in the old
grid that one of them is selected twice for the new grid (leading to a
not increasing assertion).
Assertion in abs_lookupAdapt reported by Jana.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@10782 aaf1aab0-4228-0410-ad68-8dceda47f409