* arts-2-3-1076
Starting to fix a bug in the Zeeman calculations.  It seems
the eta angle was wrong for when theta angle was close to 90 degrees.
This affects very special linear polarization calculations Luckily,
none of the use-cases I have had for this has used the geometry in
question, and the effect only redistributed some strength from 45 
to main axis polarization.  Note that this code only adds the 
necessary code to fix the problem and does not finalize the switch
to the new angles.  Expect the other update shortly.

* src/quantum.h:  Adding parity quantum number

* src/zeeman.{cc,h}:  Added separate implementations returning
angles and derivatives



git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@11029 aaf1aab0-4228-0410-ad68-8dceda47f409