* arts-2-3-1136
* src/arts.cc (arts_exit_with_error_message):

If the verbosity level for the report file is set higher than the
screen output verbosity, display a link to the report file in the error
message.

* cmake/modules/ArtsTestcases.cmake:

Change screen output level from 2 to 0. Reduces output clutter for
failed test cases. Outputs only the relevant error message, not the
whole ARTS log. The detailed log can be found in the report file if
needed.

* src/m_atmosphere.cc (AtmFieldsAndParticleBulkPropFieldFromCompact):

Remove unused species_name variable.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@11101 aaf1aab0-4228-0410-ad68-8dceda47f409