Make the code run with derivatives
This follow a very strange convention for the rest of the code.
In other parts of the code, a zero derivative is represented by
zeroes and not by an empty matrix.  I think it would be good to
make the particulate derivatives follow the more intuitive form
of derivative calculations.