Removed old optical properties extraction scheme from montecarlo.cc/h
By doing that, I also remove the workingspace method
propmat_clearskyAddParticles, because it used the old system and
there is already propmat_clearskyAddParticles2, which already uses the
new scheme. Furthermore, propmat_clearskyAddParticles2 is already used
in the test controlfile TestAbsParticle.arts.

I also removed the workingspace method TestScatDataInterp. It was apart
from propmat_clearskyAddParticles the only method/function that used the
old particle properties extraction scheme from montecarlo to check the
optical extraction of montecarlo, but montecarlo does not anymore use
the old scheme. Furthermore, it is intended for internal testing only.
With the more and more develop arts_api and the git functionality, I
don't think that we still need it.