Better abs_lookup frequency interpolation check
The combination of doing frequency jacobian together with an
absorption lookup table is quite dangerous. If the frequency
interpolation order for the table is zero, the Jacobian will be
zero, and the cause for this may be difficult for a user to
find. So we do not allow this combination.

Now there is a check in m_abs_lookup for the depreciated combination.
Furthermore, checks in m_abs_lookup are adjusted to use margins smaller
than the delta for jacobian calculation, which should be safer.