Remove old psd scheme (part 2)
All dNdD_* workingspace methods are removed and the function that
depends on them.

Due to the remove of the pnd_fieldCalcFromScat_speciesField workingspace
method in a previous commit, the workingspace variables
scat_species_*_*_field are obsolete and therefore removed. The
AtmFieldsFromCompact workingspace method is also obsolete and removed.
Test cases and internal calls are adjusted to use
AtmFieldsAndParticleBulkPropFieldFromCompact instead.

Furthermore, the scat_speciesSet is removed as it is obsolete in the new
psd scheme.

Finally, the development guide is adjusted to the new psd scheme.