Add Zeeman calculations as well
Since the Zeeman stuff is now part of the catalog, there is no longer
a need to have zeeman_linerecord_precalc.  This is thus removed from
the new version.  There will still have to be ways to modify the Zeeman
data but that is more easy to standardiz.

(Also, the Wigner library interface has to be improved.  I run into some
calls that just exits the program with no warning.  I am still wondering
how to deal with it so there is no fix for now.)