Change all make-all tests to work as they should
Most of these changes are just removing abs_lineshapes and setting
some values manually.  Some are more severe still.

One severe problem remains.  This will not be able to pass the tests
because the temporary IO I made for old catalog formats is not very
good.  It 1) does not handle .gz files, 2) does not handle
environment paths, and 3) it cannot handle the split-ARTSCAT reading
routine.  The latter I could probably fix with a few tweaks but the
former two I would have to try and understand deeper code that I wish
to avoid.