Begin adding back the line shape derivatives
My intention is to add a brief description of all the algorithms
in linefunctions.cc into this file, including the explicit derivatives
that are computed within them.  So far I have started with line shapes

The section on HTP is still not done.  It will be a complete mess
regardless of how I do it so I will just forcefully follow the
algorithm itself.