Fix very strange bug in initialization of StokesVector
For some reason, running multicore processing on the StokesVector
assignment of a Tensor4View onto a Tensor4 did not always correctly
resize the Tensor4 and delete the temporary view.  Randomly, the
error invoked was something about double free and/or corruption.
The new line added seems to fix this but makes me wonder where else
assigning a view to a matpack-target can cause a failure...